+++ R/get_assoc.R
@@ -19,13 +19,14 @@
 #' @param model A string. Default is "lm". The type of model to perform. Can also be "logistic" or "coxph"
 #' @param af Logical. Default is FALSE. Is `x` categorical? I.e., include in formula as `as.factor(x)`
 #' @param note A string. If you want to include a note like "All", "Males", "C282Y homozygotes" to describe the model or sample.
+#' @param get_fit Logical. Default is FALSE. Get model fit? (for each model: lm=R2, logistic=McFadden's pseudo-R2, coxph=Harrell's c-statistic)
+#' @param extreme_ps Logical. Default is FALSE. If p==0 then return "extreme p-values" as strings.
 #' @param scale_x Logical. Default is FALSE. Apply scale() function to exposure?
 #' @param scale_y Logical. Default is FALSE. Apply scale() function to outcome?
 #' @param inv_norm_x Logical. Default is FALSE. Apply inv_norm() function to exposure?
 #' @param inv_norm_y Logical. Default is FALSE. Apply inv_norm() function to outcome?
 #' @param subset_d Logical. Default is TRUE. Subset `d` before passing to analysis (much quicker when multiple exposures/outcomes used)
 #' @param progress Logical. Default is TRUE. Show progress bar from {purrr} `map()` function (useful when multiple exposures/outcomes provided)
-#' @param extreme_ps Logical. Default is FALSE. If p==0 then return "extreme p-values" as strings.
 #' @param ... Other `tidy_ci()` options
 #'
 #' @examples
@@ -33,12 +34,12 @@
 #' get_assoc(y="weight", x="height", z="+age+sex", d=ukb)
 #'
 #' # categorical exposure, binary outcome, and stratified analysis (with note)
-#' get_assoc(y="chd", x="smoking_status", z="+age", d=ukb |> filter(sex==1), logistic=TRUE, af=TRUE, note="Males only")
+#' get_assoc(y="chd", x="smoking_status", z="+age", d=ukb |> filter(sex==1), model="logistic", af=TRUE, note="Males only")
 #'
-#' # multiple exposures and/or outcomes
+#' # multiple exposures and/or outcomes - get pseudo R^2
 #' x_vars = c("bmi","ldl","sbp_0_avg")
 #' y_vars = c("chd","t2d")
-#' get_assoc(x=x_vars, y=y_vars, z="+age+sex", d=ukb, logistic=TRUE)
+#' get_assoc(x=x_vars, y=y_vars, z="+age+sex", d=ukb, model="logistic", get_fit=TRUE)
 #'
 #' @export
 #'
@@ -48,6 +49,7 @@ get_assoc = function(x, y, z, d,
                      model = "lm", 
                      af = FALSE, 
                      note = "", 
+                     get_fit = FALSE,
                      scale_x = FALSE, 
                      scale_y = FALSE,
                      inv_norm_x = FALSE, 
@@ -83,14 +85,18 @@ get_assoc = function(x, y, z, d,
 		d = d[,colnames(d) %in% all_vars]
 	}
 	
-	# check z formula starts with a "+" - if not, add one 
-	z = stringr::str_replace_all(z, " ", "")
-	if (stringr::str_sub(z,start=1,end=1) != "+")  z = paste0("+",z)
+	# check z formula starts with a "+" - if not, add one (unless string is empty)
+	if (z != "")  {
+		z = stringr::str_replace_all(z, " ", "")
+		if (stringr::str_sub(z,start=1,end=1) != "+")  z = paste0("+",z)
+	}
 	
 	# use {purrr} function map2() for analysis -- allows for any combination of exposure/outcome numbers
 	ret = purrr::map2(lukesRlib:::xv(x,y), 
 	                  lukesRlib:::yv(x,y), 
-	                  \(x,y) lukesRlib:::get_assoc1(x=x, y=y, z=z, d=d, model=model, af=af, note=note, scale_x=scale_x, scale_y=scale_y, inv_norm_x=inv_norm_x, inv_norm_y=inv_norm_y), 
+	                  \(x,y) lukesRlib:::get_assoc1(x=x, y=y, z=z, d=d, 
+                                                    model=model, af=af, note=note, get_fit=get_fit,
+                                                    scale_x=scale_x, scale_y=scale_y, inv_norm_x=inv_norm_x, inv_norm_y=inv_norm_y), 
 	                 .progress = progress) |> 
 	      purrr::list_rbind()
 	
@@ -104,19 +110,7 @@ get_assoc = function(x, y, z, d,
 
 
 
-#' Get tidy model output (just exposure of interest, with augmented output)
-#'
-#' @description To easily get tidy model output (from linear or logistic GLM, or CoxPH) for a categorical or continuous exposure, including sample size (and N cases if logistic), outcome, and model info. Idea is to make quick loops easy.
-#'
-#' For all exposures, it gets the N. For categorical exposures, the N is split by group, and a row is included for the reference category
-#'
-#' !!!! Better to just use `get_assoc()` though, which calls this function but after checking variables and subsetting etc.
-#'
-#' @return Returns a tibble - summary statistics from a model 
-#'
-#' @author Luke Pilling
-#'
-#' @name get_assoc1
+#' Internal function to get tidy model output (just exposure of interest, with augmented output)
 #'
 #' @param x A string. The exposure variable name, found in `d`
 #' @param y A string. The outcome variable name, found in `d` -- if model is 'coxph' then paste the survival object here e.g., 'Surv(time, event)' where `time` and `event` are variables in `d`
@@ -129,24 +123,10 @@ get_assoc = function(x, y, z, d,
 #' @param scale_y Logical. Default is FALSE. Apple scale() function to outcome?
 #' @param inv_norm_x Logical. Default is FALSE. Apple inv_norm() function to exposure?
 #' @param inv_norm_y Logical. Default is FALSE. Apple inv_norm() function to outcome?
+#' @param get_fit Logical. Default is FALSE. Get model fit? (lm=R2, logistic=McFadden's pseudo-R2, coxph=Harrell's c-statistic)
 #' @param ... Other `tidy_ci()` options
 #'
-#' @examples
-#' # for one outcome, equivalent to `tidy_ci(glm(weight ~ height +age+sex, d=ukb))` - with added `n`
-#' get_assoc1(y="weight", x="height", z="+age+sex", d=ukb)
-#'
-#' # categorical exposure, binary outcome, and stratified analysis (with note)
-#' get_assoc1(y="chd", x="smoking_status", z="+age", d=ukb |> filter(sex==1), logistic=TRUE, af=TRUE, note="Males only")
-#'
-#' # multiple exposures on single outcome, then combine output
-#' x_vars = c("bmi","ldl","sbp_0_avg")
-#' res = do.call(rbind, lapply(x_vars, get_assoc1, y="chd", z="+age+sex", d=ukb, logistic=TRUE))
-#'
-#' # one exposure on multiple outcomes, then combine output
-#' y_vars = c("bmi","ldl","sbp_0_avg")
-#' res = do.call(rbind, lapply(y_vars, function(y) get_assoc1(x="smoking_status", y=y, z="+age+sex", d=ukb, af=TRUE)))
-#'
-
+#' @noRd
 get_assoc1 = function(x, y, z, d, 
                       model = "lm", 
                       af = FALSE, 
@@ -155,14 +135,8 @@ get_assoc1 = function(x, y, z, d,
                       scale_y = FALSE,
                       inv_norm_x = FALSE, 
                       inv_norm_y = FALSE,
+                      get_fit = FALSE,
                       ...)  {
-
-	# check inputs
-	if (class(x) != "character")  stop("x needs to be a string, a variable name in d")
-	if (class(y) != "character")  stop("y needs to be a string, a variable name in d")
-	if (class(z) != "character")  stop("z needs to be a string of variable names in d")
-	if (! any(class(d) %in% c("data.frame","tbl","tbl_df")))  stop("d needs to be a data.frame or tibble")
-	if (! any(model %in% c("lm","logistic","coxph")))  stop("model needs to be 'lm' 'logistic' or 'coxph'")
 
 	# what model are we doing?
 	logistic = FALSE
@@ -183,10 +157,6 @@ get_assoc1 = function(x, y, z, d,
 			d = d |> dplyr::mutate(!! rlang::sym(y1) := as.numeric(!! rlang::sym(y1)) )
 		}
 	}
-
-	# check variables are all in d
-	if (! x %in% colnames(d))  stop("x needs to be a variable in d")
-	if (any(! yy %in% colnames(d)))  stop("y needs to be a variable in d")
 
 	# exposure variable - categorical?
 	xx = x
@@ -283,6 +253,32 @@ get_assoc1 = function(x, y, z, d,
 			res = res |> dplyr::mutate(n, n_cases)
 		}
 
+	}
+
+	# get model fit?
+	if (get_fit)  {
+		fit_stat = NA
+		if (!logistic & !coxph)  {
+			# linear model - get R^2
+			fit_stat = cor(fit$y,predict(fit))^2
+		}
+		if (logistic)  {
+			# logistic model - get pseudo-R^2 (1 - log(likelihood) / log(likelihood - null model))
+			#  - need to make sure null model has same N has full model - exclude any missing X or Z
+			z_vars = stringr::str_replace_all(z, " |as.factor\\(|\\)", "") |> 
+			         stringr::str_split_1(stringr::fixed("+")) |> 
+			         purrr::keep(\(x) stringr::str_length(x)>0)
+			if (length(z_vars)==0)  z_vars = yy
+			fit_null = glm(paste0(yy, " ~ 1"), data=d |> select(yy, x, z_vars) |> na.omit(), family=binomial(link="logit"))
+			fit_stat = 1-logLik(fit)/logLik(fit_null)
+			fit_stat = fit_stat[1]
+		}
+		if (coxph)  {
+			# CoxPH model - get C-statistic
+			fit_stat = concordance(fit)
+			fit_stat = fit_stat$concordance
+		}
+		res = res |> dplyr::mutate(fit_stat)
 	}
 
 	# modify final bits